requirements.txt
alembic==1.5.4
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1

# alembic==1.7.4
# attrs==21.2.0
# click==8.0.3
# Flask==2.0.2
# Flask-Migrate==3.1.0
# Flask-SQLAlchemy==2.5.1
# iniconfig==1.1.1
# itsdangerous==2.0.1
# Jinja2==3.0.2
# Mako==1.1.5
# MarkupSafe==2.0.1
# packaging==21.0
# pluggy==1.0.0
# psycopg2-binary==2.9.1
# py==1.10.0
# pyparsing==3.0.3
# pytest==6.2.5
# python-dotenv==0.19.1
# SQLAlchemy==1.4.26
# toml==0.10.2
# Werkzeug==2.0.2
